Updates: nimble csv, breaking out batch_upsert and from_parts functions
