+++ lib/draft/pick_data_setup/parsing_helpers.ex
